Refactor JS files to TypeScript and update ESLint rules
